functions/kit/initiate_upload/requirements.txt
boto3
pyhumps
ujson